+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.1.9] - Unreleased
+
+### Changes
+
+- Added path for "nonce" and "state" cookies
+- Added cookies cleanup on successful login
+
+
 ## [1.1.4] - 2017-01-05
 
 ### Changes